requirements.txt
Werkzeug==3.1.4
bcrypt==5.0.0
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
Flask==3.0.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
openai==1.35.10
passlib==1.7.4
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.1